Refactor: detail파일 위치 이동
