+++ src/pages/detail/Detail.tsx
